requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
motor>=3.5.0
beanie>=1.25.0
pydantic==2.6.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
fastapi-mail==1.4.1
email-validator==2.1.0.post1
python-dotenv==1.0.1
